pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lgtm-mcp"
version = "0.1.0"
description = "MCP server for Loki, Prometheus, and Tempo APIs"
requires-python = ">=3.11"
license = "MIT"
readme = "README.md"
keywords = ["mcp", "loki", "prometheus", "tempo", "grafana", "observability"]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-PyYAML",
]

[project.scripts]
lgtm-mcp = "lgtm_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lgtm_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["lgtm_mcp"]

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
warn_return_any = false
warn_unused_ignores = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
